+++ src/backend/app/models/__init__.py
@@ -4,4 +4,4 @@ from .user import User, UserManager
 from .project import Project, ProjectFile, ProjectLanguage, ProjectFramework, Contributor, ProjectContribution
 from .resume import Resume
 from .portfolio import Portfolio, PortfolioProject
-from ..models import ProgrammingLanguage, Framework
+from .language import ProgrammingLanguage, Framework
